+++ src/Component/OpenApiCommon/Naming/OperationIdNaming.php
@@ -2,6 +2,7 @@
 
 namespace Jane\Component\OpenApiCommon\Naming;
 
+use Jane\Component\JsonSchema\Generator\Naming;
 use Jane\Component\JsonSchema\Tools\InflectorTrait;
 use Jane\Component\OpenApiCommon\Guesser\Guess\OperationGuess;
 use Symfony\Component\String\Slugger\AsciiSlugger;
@@ -12,10 +13,12 @@ class OperationIdNaming implements OperationNamingInterface
     use InflectorTrait;
 
     private SluggerInterface $slugger;
+    private Naming $naming;
 
     public function __construct()
     {
         $this->slugger = new AsciiSlugger();
+        $this->naming = new Naming();
     }
 
     public function getFunctionName(OperationGuess $operation): string
@@ -28,6 +31,6 @@ class OperationIdNaming implements OperationNamingInterface
         $operationId = (string) $operation->getOperation()->getOperationId();
         $operationId = $this->slugger->slug($operationId, '-');
 
-        return $this->getInflector()->classify($operationId);
+        return $this->naming->fixReservedClassName($this->getInflector()->classify($operationId));
     }
 }